Add BGB implementation
Now, I know this repository says "hardware", but hear me out:
libmobile is made to be easily portable across a plethora of devices.
Now, faced with the inability to test the Mobile Trainer GB without the
help of an emulator, this seemed like a really good candidate for a port.

I'm a bit unhappy with the poor performance delivered by this on linux
(haven't tested this on windows proper yet, only under WINE). I've tried
a plethora of different things to improve it, but it doesn't seem to
want to work. For this reason, I might try porting it to mGBA, next.
Who knows.