!!! Split repository and relicense
This repository has been split off from the original ArduinoAdapterGB
repo using the following command:
    git-filter-repo --path libmobile --path-rename libmobile/:

This is to facilitate easier use of the library by vendoring or
including this repository as a submodule, as well as allow independent
development of the library from the implementations.

I hereby declare that all of the code was originally copyrighted by me
(commit author), and as such can be relicensed.