WIP: Implement p2p relay protocol
This protocol will be used to relay peer-to-peer connections in a way
that's semi-transparent to the user. This effectively solves the issue
of having to bypass the user's NAT somehow, by just forwarding it all
through a central server. With the added bonus of being able to assign
individual "phone numbers" to the users.

Of course, a solution like this isn't very scalable, but these games
transfer very little data, on very slow machines, so it is serviceable.
If we end up with too many users, it isn't too hard to tell users to
just hop servers, or implement some sort of multi-server load balancing.
But I honestly doubt that'll be necessary.