WIP: Refactor callbacks to allow using pointers
This change is one in a few that will allow this library to expose a
stable ABI when used as a shared library. This also fixes the fact that
windows doesn't support weak symbols, and every application would need
to make sure to export all of the weak symbols, preventing us from
easily adding new ones without updating every windows application using
it.

Meanwhile, this functionality can be toggled for implementations where
the memory and performance gains are significant enough to warrant
throwing away the pointers.